Created handlers for methods. Removed unused classes.
Added methods for getting interests and interestlinks for a single person
