--- Labb3-miniAPI/Models/DTO/PersonDTO.cs
@@ -1,8 +1,0 @@
-﻿namespace Labb3_miniAPI.Models.DTO
-{
-	public class PersonDTO
-	{
-		public string Name { get; set; }
-		public string Phone { get; set; }
-	}
-}
