+++ Labb3-miniAPI/Program.cs
@@ -1,4 +1,5 @@
 using Labb3_miniAPI.Data;
+using Labb3_miniAPI.Handlers;
 using Labb3_miniAPI.Models;
 using Microsoft.EntityFrameworkCore;
 
@@ -16,26 +17,29 @@ namespace Labb3_miniAPI
 			app.MapGet("/", () => "Hello World!");
 
 			//POSTS
-
+			
 			//adds a new person to the database
-			app.MapPost("/person", Person.AddPerson);
+			app.MapPost("/person", PersonHandler.AddPerson);
 
 			//Adds a new interest to a person
-			app.MapPost("/person/{id}/interest", Interest.AddInterestToPerson);
+			app.MapPost("/person/{id}/interest", InterestHandler.AddInterestToPerson);
 
 			//Adds a link to a specic interest and person
-			app.MapPost("/person/{personId}/interest/{interestId}/interestlink", InterestLink.AddLinkToInterestAndPerson);
+			app.MapPost("/person/{personId}/interest/{interestId}/interestlink", InterestLinkHandler.AddLinkToInterestAndPerson);
 
 			//Connects an existing person with an existing interest and, if available, corresponding links. Based on ID for person and interest
-			app.MapPost("/person/{personId}/interest/{interestId}", Person.AddPersonToExistingInterest);
+			app.MapPost("/person/{personId}/interest/{interestId}", PersonHandler.AddPersonToExistingInterest);
 
 			//GETS
 
 			//Lists all people in the database with name and id
-			app.MapGet("/person/all", Person.ListAllPeople);
+			app.MapGet("/person/all", PersonHandler.ListAllPeople);
 
-			//Lists a specific person with their interests and interestlinks
-			app.MapGet("/person/{id}", Person.ListSpecificPerson);
+			//Lists a specific persons interests
+			app.MapGet("/person/{id}/interests", InterestHandler.ListInterestsForId);
+
+			//Lists a specific persons links
+			app.MapGet("/person/{id}/links", InterestLinkHandler.ListInterestLinksForId);
 
 			app.Run();
 		}
